feat general: add "Natural Language" section
The clause stating that English language should be used for all names
and comments had been removed from the "General" section of the C++
guide. In turn, the general guidelines gained a section called "Natural
Language" which establishes English as the only official language for
the whole community, but permits Russian in private talks with Igor Gorlov.